chore(admin-ui):cleanup
